Learning_Platform_Analysis/requirements.txt
mysql-connector-python==8.2.0
pandas==2.1.4
numpy==1.24.3
openpyxl==3.1.2